+++ bbt/tests/server/evaluate_tck/invalid_input_type/run.sh
@@ -23,7 +23,7 @@ rm data.json
 curl -s -X POST http://0.0.0.0:22022/definitions/deploy > /dev/null 2>&1
 
 # evaluate model using generic input data format
-curl -s -d '{"model":"compliance-level-2-test-0001","invocable":"Greeting Message","input":[{"name":"Full Name","value":{"simple":{"type":"xsd:tekst","text":"John Doe","isNil":false}}}]}' -H "Content-Type: application/json" -X POST http://0.0.0.0:22022/tck/evaluate
+curl -s -d '{"namespace":"io.dmntk","model":"compliance-level-2-test-0001","invocable":"Greeting Message","input":[{"name":"Full Name","value":{"simple":{"type":"xsd:tekst","text":"John Doe","isNil":false}}}]}' -H "Content-Type: application/json" -X POST http://0.0.0.0:22022/tck/evaluate
 
 # stop dmntk server
 kill -s SIGINT "$_pid"