+++ workspace/src/errors.rs
@@ -48,10 +48,10 @@ pub fn err_model_evaluator_is_not_deployed(s: &str) -> DmntkError {
   WorkspaceError(format!("model evaluator for definitions '{s}' is not deployed")).into()
 }
 
-pub fn err_definitions_with_namespace_already_exists(s: &str) -> DmntkError {
-  WorkspaceError(format!("definitions with namespace '{s}' already exist in workspace")).into()
+pub fn err_invalid_namespace(s: &str) -> DmntkError {
+  WorkspaceError(format!("invalid namespace '{s}'")).into()
 }
 
-pub fn err_definitions_with_name_already_exists(s: &str) -> DmntkError {
-  WorkspaceError(format!("definitions with name '{s}' already exist in workspace")).into()
+pub fn err_definitions_with_name_already_exists(namespace: &str, name: &str) -> DmntkError {
+  WorkspaceError(format!("definitions with name '{name}' already exist in namespace '{namespace}'")).into()
 }